Add: useGoogleAnalytics hook 으로 분리
