+++ next-app/src/hooks/useGoogleAnalytics.ts
@@ -1,0 +1,22 @@
+import { useEffect } from "react";
+import { useRouter } from "next/router";
+import * as gtag from '../utils/gtag';
+
+
+const useGoogleAnalytics = () => {
+  const router = useRouter();
+
+  useEffect(() => {
+    const handleRouteChange = (url: string) => {
+      gtag.pageview(url);
+    };
+    router.events.on('routeChangeComplete', handleRouteChange);
+    router.events.on('hashChangeComplete', handleRouteChange);
+    return () => {
+      router.events.off('routeChangeComplete', handleRouteChange);
+      router.events.off('hashChangeComplete', handleRouteChange);
+    };
+  }, [router.events]);
+}
+
+export default useGoogleAnalytics;
